wallet-account.ts separated to several files
