+++ lib/common-index.d.ts
@@ -11,5 +11,5 @@ import { Signer, InMemorySigner } from './signer';
 import { Contract } from './contract';
 import { KeyPair } from './utils/key_pair';
 import { Near } from './near';
-import { ConnectedWalletAccountRedirect, WalletConnectionRedirect } from './wallet-account';
+import { ConnectedWalletAccountRedirect, WalletConnectionRedirect } from './wallet/index';
 export { accountCreator, providers, utils, transactions, validators, multisig, Account, Connection, Contract, InMemorySigner, Signer, KeyPair, Near, ConnectedWalletAccountRedirect, WalletConnectionRedirect };
