+++ lib/common-index.js
@@ -46,6 +46,6 @@ const key_pair_1 = require("./utils/key_pair");
 Object.defineProperty(exports, "KeyPair", { enumerable: true, get: function () { return key_pair_1.KeyPair; } });
 const near_1 = require("./near");
 Object.defineProperty(exports, "Near", { enumerable: true, get: function () { return near_1.Near; } });
-const wallet_account_1 = require("./wallet-account");
-Object.defineProperty(exports, "ConnectedWalletAccountRedirect", { enumerable: true, get: function () { return wallet_account_1.ConnectedWalletAccountRedirect; } });
-Object.defineProperty(exports, "WalletConnectionRedirect", { enumerable: true, get: function () { return wallet_account_1.WalletConnectionRedirect; } });
+const index_1 = require("./wallet/index");
+Object.defineProperty(exports, "ConnectedWalletAccountRedirect", { enumerable: true, get: function () { return index_1.ConnectedWalletAccountRedirect; } });
+Object.defineProperty(exports, "WalletConnectionRedirect", { enumerable: true, get: function () { return index_1.WalletConnectionRedirect; } });
