+++ lib/wallet/injected/wallet-injected-interface.js
@@ -1,0 +1,2 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
